third_party/WebKit/LayoutTests/fast/events/zoom-dblclick-expected.txt



Zoomed
PASS event.clientX is within 0.00001 of 166.66667
PASS event.clientY is within 0.00001 of 83.33333
PASS event.pageX is within 0.00001 of 166.66667
PASS event.pageY is within 0.00001 of 83.33333
PASS successfullyParsed is true

TEST COMPLETE
